+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.2
+
+* Support Location Permission
+
+## 0.0.1+1
+
+* Migrate To AndroidX
+
 ## 0.0.1
 
 * Initial Release